[RPP] Show resolved name for profile calls in their details
Trace event details that correspond to a JS function have an entry detailing the function name and location. If possible, this data should be mapped when source maps are available, something that wasn't being done for profile calls of the new engine.

Took the chance to introduce a helper with the boilerplate to register a sourcemap in the unit test environments, and used that in a new test and  in an existing one.

Bug: none
Change-Id: I2cd8730b31d29aaf0c76d0917bbab389017e1675