[settings] Properly ellipsize long ignore list patterns.
The <label> for the checkboxes already had `text-overflow: ellipsize` on
it, however that is not effective unless the container of that element
has `overflow: hidden` and `white-space: nowrap`, and the width of the
container element is also set accordingly. This CL fixes both (the
latter only for the case of the exclusion rules).

Fixed: chromium:1377254
Before: https://imgur.com/tquPmFW.png
After: https://imgur.com/8vnfNJC.png
Change-Id: Ie1f96be5109d1c3475d0be8576c9c75969a64e60