[RPP] Remove indexForEvent map
The map is used to allow for appenders to modify flame chart entry data
after an entry has been appended using the corresponding index in the
flame chart entry data . However given the size of the data  stored in
it (an entry for each event/profile call) adding data to it eventually
becomes very expensive.

Turns out we don't really need it. In this CL, a hook is added to the
helper that appends the flame chart data which is called back each time
an entry is added. This way, appenders have access to the entry's index
without having to incur in the extra cost.

Before: https://screenshot.googleplex.com/zRmccvmW5ZV83Vu
After: https://screenshot.googleplex.com/6P6NEUFK2pnzLiB

Bug: 1498918
Change-Id: Iae6b2943b59527ebd8c5e916b82ef6cdc3f4067d