[RPP] Show resolved name for profile calls in their details
Trace event details that correspond to a JS function have an entry detailing the function name and location. If possible, the function name should be mapped when source maps are available, something that wasn't being done for profile calls of the new engine.

Took the chance to introduce a helper with the boilerplate to register a sourcemap in the unit test environments, and used that in a new test and  in an existing one.

Before: https://screenshot.googleplex.com/BRSNQAEguH9YVQ3
After: https://screenshot.googleplex.com/B6yCgaNGMEKzA9U

Bug: none
Change-Id: I2cd8730b31d29aaf0c76d0917bbab389017e1675