[RPP] Do not redraw minimap when data is unchanged
Drawing the minimap is expensive. At the moment we are doing it multiple
times at trace load without necessity. This is a first step to avoid
unnecessary work. If new data is set but it's equivalent to the last
drawn data (f.e. because names were resolved with source maps) don't
redraw.

Note from the screenshots that we are drawing the minimap 4 times. After
this CL we are drawing it 3 times. It ain't much but it's honest work.
Further improvements will be done as follow ups.

Before: https://screenshot.googleplex.com/6eDLsGPLoUau3p2
After: https://screenshot.googleplex.com/5TFBsYurXnVt3Nh

Bug: 1492405
Change-Id: I17bf7c236249d00d7b98cb4362816713b2dca293