RPP: Remove ThreadTracksSource.BOTH_ENGINEs and update screenshots
This is the first in a long set of CLs that will begin to remove old
engine functionality we do not need.

I have started with the ThreadTracksSource enum and removed the
BOTH_ENGINES flag; updating all the screenshot tests to use the new
engine exclusively. In future CLs I will remove the enum and the flag
entirely, but first I will go through and remove code that only runs for
the OLD_ENGINE flag.

Bug: 1499182
Change-Id: I7ab164e500576d044f334705d5e99649b3523e94