[RPP] Reduce expensive calls in pie chart data building
Building the piechart implies querying each event's category, which
depending on how it's done can get expensive. This CLs tweaks the
querying for categories so that the overall cost is reduced. Also, the
check for event's style visibility is removed since all events passed to
the builder function are coming straight from the flame chart (and as
such are visible).

In the feature we could further improve this if we calculated the timings as we are building the trace entry tree in the trace engine

Before: https://screenshot.googleplex.com/6LYi7cBgUtq2nsd
After: https://screenshot.googleplex.com/nKi3X7kbyibPyAF

Bug: 1472155
Change-Id: I4b0cdaa36049fbf00d4a3692553fd3fe757d6bf6