[sources] Add indentation markers to DevTools.
This is similar to the indentation markers feature found in VS code. We
add a temporary experiment to disable the feature and will track usage
of that experiment to decide on whether we need an option in the future.

Screenshot: https://imgur.com/AaFWzdH.png
Fixed: chromium:1479986
Doc: http://go/chrome-devtools:indentation-markers-proposal
Change-Id: I483af8517f1ead3507d5fabacad085b698c4578b