RPP: land tooltip for interaction breakdown
This CL adds breakdowns to interaction entries when they are hovered
over in the timeline.

As a drive-by it also renames "Main thread processing" to "Processing
time", because interactions could come from other frames/threads. I
updated the key in the TimelineUIUtils details view too so the tooltip
and the details drawer are consistent.

Bug: 1495751
Change-Id: I9e63d7e82f48de76cfcf71b65bb2bb4dd2c9dfdd