[Watch build] Enable running unit tests on file change
The part that takes a lot of time in running unit tests is actually
spinning up the browser and connecting to it.

`watch-unittests` command starts the file watcher for Karma
test runner which runs the tests on file change.

You can use this without `watch` script and by just rebuilding
the target.

When combined with `watch` command, this enables you to run unit tests
instantly after a file change.

Bug: none
Change-Id: Icd71230b5b455254dc25b4cef5f6ae72c5dbba06