[Protocol Monitor] Improve performance of protocol monitor and select cell
Updates are:
* Debounce calls to `data` setter in `update`
* Do not copy & save messages in Protocol Monitor for exporting to file logic
* Select the request/response exactly based on the selected cell

For some time, trying to debug an iframe issue; I was having problems with
issue not appearing when the protocol monitor is open and a filter is enabled.

In addition to that, refreshes to a page while the protocol monitor is open
and a filter is active was unbearably slow.

This CL fixes the performance issue for that case with 2 improvements.

As a drive by: updated the tab selection logic to select the tab
based on the focused cell of the protocol message.

Bug: none
Change-Id: I8819db5dc47e14ac32fb47e5ea7621a4cf4fb7b5