[elements] Switch Elements' sidebars to ToolbarInput.
This changes Styles, Computed, and Properties to use the ToolbarInput
component instead of handcrafting input fields here and making them
look and behave like the other filter fields in DevTools.

Drive-by-fix: Also simplify the `filterTextField` VisualElement to just
`textField` and use that consistently for ToolbarInput. The surrounding
context of the toolbar makes it easy to distinguish what the textField
does, and if we need more distinction, we can always introduce optional
context and thread that through.

Fixed: chromium:1457416
Bug: chromium:1467464
Change-Id: Iaf2dd52d825c72ee34e1767b98236d19cd0308e7