[developer_resources] Add visual element logging.
This CL instruments the Developer Resources panel for visual element
logging. In doing so, it also introduces (initial) support for logging
to the legacy DataGrid.

Drive-by-fix: Change the ID of the panel to 'developer-resources' to be
consistent in the logging (as the ID is also used for the context).

Fixed: b:308381558
Bug: chromium:1467464
Change-Id: I906ce4089fae729ebccabe487240cbfc26859b4c
Screenshot: https://imgur.com/jxaxfC9.png