[RPP]: Layer Tree Handler
As part of replacing the old engine, we need to replace the
EventTimelineData structure which holds a bunch of info about events.

One of those fields is the `picture` event. This ties a snapshot event
to a given Paint event.

NOTE: these snapshots are only available with "Advanced Paint
Instrumentation" turned on. The frames track does NOT show these, it
uses "Screenshot" events, which are always available.

This CL creates a new LayerTreeHandler which copies the logic in the
TimelineModel for tracking various events in order to associate these
snapshots with paints. By landing this, we can then remove any instance
where we look up `EventTimelineData.forEvent(event).picture` with a call
into this handler.

It is only ever Paint events that get given a screenshot, so we can
enforce that via the new types that this CL includes.

Bug: 1499182, 1478742
Change-Id: Ifdcf570d62f49bb0b750357385e8d2bfbadaacfe