Fix hover style for data grid
Previously, the hover style would apply a background color that is
mostly transparent, with the intention to slightly darken the original
color. This sometimes does not work because it replaces the original
color instead of overlaying it.

The solution is to move the original color of the column header to the
parent tr element.

R=kprokopenko@chromium.org

Fixed: 1502390
Change-Id: I09ebec30eedce4147497ac761aae697d9a309d6d