[cleanup] Use `UI.ActionRegistraction.Action` more consistently for VEs.
This is a follow-up to https://crrev.com/c/5033661 to make use of the
existing `Action` machinery more consistently for VE actions as well.
This has the additional benefit that users will also be able to use
these actions from the command menu as well as by clicking on the
relevant button (and we track both activations already via UMA).

For the case where it's not easily possible to use `Action`s, in
`SearchView`, we now at least already prefix the VE context's
appropriately so that at some later stage we might be able to
use the default machinery here. And while we are here, also fix
a copy&paste bug in the reporting for the clear button.

Bug: chromium:1467464, b:308381657, b:308382370
Change-Id: I76ef66809008306d9610a838dc14a324e7cb1c17