[CSSStyleDeclaration] Remove parseUnusedText logic
It is introduced in crrev.com/c/1089295 and is served well to us
until now, with the nesting introduced.

Its custom logic for trying to figure out which properties
might the user intended to write is getting more and more complex with
the addition of new CSS features (nesting again...).

So, this is getting a risky logic that would break Styles pane
in some situations.

The UX win it provides (crbug.com/813435) does not justify
our maintenance costs and the risk it introduces to the reliability.

So, it's time we say goodbye to our fellow logic.

Fixed: 1473339
Change-Id: Ieaf87f7f9dbcc02d835ec042628a75ec0e13de15