[RPP]: migrate TimelineFrameModel to new engine
This CL introduces (but does not use in production) a new
TimelineFrameModel. Rather than rework the logic, it is a copy of the
existing model but updated to use the new trace engine data and types.

The TimelineFrameModel is a large blocker towards removing the old
engine hence why it needs to be removed and replaced by this one.

The next steps after landing this CL are:

1. Work to remove the dependency on SDK: this CL adds a dependency to
   SDK for the models/trace folder, which we do not want given future
   ambitions to reuse this code outside of DevTools.
2. Migrate the Frames track to one powered by this new
   TimelineFrameModel.
3. Remove the old TimelineFrameModel.

Change-Id: I132a5ea7ad20a05a9810af01c8e6dbacb6944714
Bug: 1499182