[cleanup] Don't use nested panels for CSS Overview.
This is a follow-up to https://crrev.com/c/5037114, which added
instrumentation to CSS Overview. But the instrumentation revealed
that the CSS Overview was nesting a UI.Panel.Panel inside of another
UI.Panel.Panel, which leads to very weird instrumentation and is
not how we generally envision to use panels.

This CL fixes that and simply puts a UI.Widget.VBox combined with
UI.SplitWidget.SplitWidget in place.

Bug: b:311314000, chromium:1467464
Change-Id: I6bce2348e74113ece724281c07b8c14a33847160