[TextUtils] Allow regex filters to end with escaped characters.
The fix from https://crrev.com/c/4768202 made the regular expression we
use to identify filter regex's too strict and didn't allow for regex's
to end with escaped characters, like /search\?/ and the like. This fix
lifts that restriction and also accepts regex's that end in escaped
characters.

Bug: chromium:1346936
Change-Id: I00cc82a6bd6cd4564e81d91a8f22436ae1a86b68
Fixed: chromium:1502570