[infra] Fix downloading emscripten binaries on mac-arm64
This differentiates the url paths for mac-arm64 emscripten
binaries. Since some bots already downloaded the wrong x64
binaries we need to clobber them once. In order to do so, this
change also changes the stamp-file content to the full URL,
which will cause a re-download of the binaries once (for
everyone).

Cq-Include-Trybots: luci.devtools-frontend.try:devtools_frontend_mac_arm64_rel
Bug: chromium:1500139
Change-Id: I004b45614b1a54ce5fe6ab871bb69f4575a059e8