[sources] Fix execution token highlighting.
CodeMirror6 constructs the syntax tree on-demand, and therefore might
return an incomplete syntax tree when stopping in a newly opened file
and trying to compute the execution token decoration. The fix here is
to give CodeMirror6 parser some time to finish the contruction of the
syntax tree at least up to the execution line.

Fixed: chromium:1496475
Change-Id: Ib2f2afbb5ffb2cd81de0d98b7609c06524c58d3b