Move LayoutShift initiators out of RootCauses
The logic here was duplicated (and slightly different) from the
InitiatorsHandler which is already doing most of the work required. This
CL updates the root causes code to use the initiators handler, rather
than re-implement initiators.

This has the nice side-effect that we can remove the requirement to
calculate initiators as a by-product of TreeHelpers.treify.

Bug: 1434596
Change-Id: Ifdae8f53e6e65c43e4aca9ff24c3651dccb7a0b8