[sources] Clamp issue source locations to the document.
Chrome sometimes reports out of document source locations, and we
already deal with out of document column numbers, but so far we did not
properly deal with out of document line numbers, which are especially
problematic with automatic pretty printing.

This CL also clamps line numbers and re-enables the relevant test that
flushed out the bug.

Bug: chromium:1503031, chromium:1382752
Change-Id: Idcabc48e4353cdf53d141b31562ac9aca4e2ab0a