RPP: change Node.parentId to Node.parent
I have a situation where, given a node, I need to walk up the tree to
find its parent (& then keep walking). With `parentId` this is hard; we
do not expose from the handler data that would enable a caller of the
trace engine to do this.

To make this easier, and also to make it consistent with how we define
`children` to be an array of nodes (not of IDs), this CL changes from
`parentId` to `parent`.

Bug: none
Change-Id: Ib219d4ff9e5027dafa3361667a28a398c065e3c3