RPP: draw initiator arrows with new engine
This CL migrates the flame chart's initiator arrow drawing to use
initators defined in the new engine.

There are still some places in the code (in TimelineUIUtils) where we
rely on the old initiator logic, primarily around network requests, and
around showing stack traces for events. Those need a bit more
investigation and thought, so I have resisted the urge to rip them out
as part of this work! I will continue to investigate in follow-ups.

Bug: 1434596
Change-Id: Id21ddb7e806226a04d4b707d7a9608b62a476281