[build] Fix front_end/panels/recorder/injected build config
`injected` depends on `third_party/{puppeteer, puppeteer-replay}`.
And their outputs are in `gen` dir.
But, the source files of `injected` aren't in `gen` dir.
So, it fails to find the generated files.

This CL addresses this issue by copying the source files to gen, and build `devtools_module` under the `gen` dir.

Bug: b/308405411
Change-Id: If05fb9d188f8c1450a2a85a70857b4a11e4af349