Remove legacy JS Profiler model parsing
This CL trims down `TimelineJSProfile` into the few functions that we
actually use now that all the parsing is done in SamplesHandler by the
new engine.

In time we should probably move the remaining functions here into the
trace engine itself, but let's leave them for now - once all the old
dead code is removed this might become clearer.

Bug: 1499182
Change-Id: Ia0262f51c96eb1395e4eb36bcdbc556da1d01657