Remove TimelineFrameModel and migrate to FramesHandler
IMPORTANT: do not merge before the branch cut on Dec 4.

This CL removes the old TimelineFrameModel and migrates all usages of it
to the new engine.

Unfortunately this is a rather large CL, but there was no real way to
incrementally do this wwork without maintaining both frame models (old
and new) and implementing some way of linking them which did not seem
worth it.

This CL updates the frames track, paint profiler and layers view to use
data from the new frames model. Mostly this means implementing SDK based
lookups that were in the legacy TimelineFrameModel in other parts of the
codebase (to ensure the Trace engine does not rely on SDK).

I have manually tested this locally on a few different traces, and have
confirmed that things seem to be behaving as expected. That said, it
seems silly to risk landing this now just before a branch point, so I
will wait until next week.

There are a few TODOs left in the code; this CL was already big so I
have left those on purpose and will follow this CL up with those
improvements.

I also removed a "Show layers" link that is shown in the details view if
the frame has layers; this link is actually broken on stable right now,
and I wasn't entirely sure how to implement it. Given that, I have
removed the link (you don't need it, as the Layers tab appears). We can
consider if to bring the link back again after this CL lands.

Bug: 1499182
Change-Id: Id881202b944f814cbae0c45d9ffb0a7032030701