[ui] Use less confusing icon for 'Collect garbage'.
Prior to this CL, the 'Collect garbage' action in the Performance and
Memory panels would use the Bin icon from Material 3, but that leads
to a lot of confusion, since that icon usually means 'Delete' or 'Move
to Trash' in all other contexts (which for example when looking at the
screenshot below, one would usually think that the Bin button would
remove the selected profile from the list in case of the Memory
panel).

With this CL we now use the Mop icon from Material 3, which is not
used often in other contexts and is still sufficiently closely related
to collecting garbage.

Fixed: chromium:1507530
Screenshot: https://imgur.com/qVv6tGE.png
Change-Id: I309d8b6b403918cd285087c98d6d532984a5d054