[cleanup] Make UI.ContextMenu.Provider type-safe and remove singletons.
Similar to https://crrev.com/c/5054330, this introduces a type parameter
for the context menu providers, which allows for better type safety and
more importantly clearly documents the contract.

Also remove the useless singletons for UI.ContextMenu.Provider
subclasses, as these are cheap to construct and have no identity (except
for some panels, where we of course need to keep the singletons), and
they are much cheaper to construct on-the-fly than to maintain the
singletons.

Bug: chromium:1379221
Change-Id: If6f2bc4887bb73f9dd70a4c5e52c45c355257d2d