Add milliseconds to TraceBounds
Today I began prototyping what it would look like to use the TraceBounds
service as the source of truth for all timings within the Performance
Panel.

I discovered that it would be nice if the service dispatched millisecond
values also (as lots of the UI uses milliseconds), and that we should
just dispatch one event when any data changes, rather than one-per field
that changes. This means when all of them change we can dispatch one
event, rather than three which would cause potentially three re-renders.

As part of the work to remove the old engine, I am working through the
PerformanceModel's code around windows and time changes, and so next
after this CL lands I will be attempting to migrate the UI to use the
TraceBound service.

Bug: 1499182, 1491689
Change-Id: Id378daeed430e21cdd59d391cba46597e93b7c2a