Use `typeRoots` for custom type-defs
crrev.com/c/5081999 added in custom WebGPU types, however doing so via
`types:[]` overrides the TS default behaviour to include types from
`node_modules/@types`. This means in VSCode/LSP based editors, we now
get TS errors in tests from `assert` and `sinon`, which TS now cannot
detect (as it is not reading `node_modules/@types/...`).

To prevent this we can instead use `typeRoots`, and specify the folders
to use. This maintains the default behaviour and adds detection of the
WebGPU types.

Bug: 1506431
Change-Id: Ie5aed47bc630852305ecc101d05d7081dfa71647