[cleanup] Introduce a `LinearMemoryInspectable` revealable.
This adds add revealable named `LinearMemoryInspectable`, which is
essentially a pair of `SDK.RemoteObject.RemoteObject` (that is linear
memory inspectable) and an optional `expression`. This is the first
step to break the bogus dependency of the `object_ui` module on the
`linear_memory_inspector` module, which prevents the latter from
being a proper panel module.

Bug: chromium:1507747
Change-Id: I196583cbd63c308f54e89fe412d1c10aed17f935