RPP: fix details view for V8.Compile events
V8 emits two events for compiling scripts:
- V8.CompileScript
- v8.compile (yes, it is all lowercase for some unknown reason!)

In the old engine, we used the v8.compile event to represent the script
compilation because it is the event that contains the metadata
associated with the compilation such as the script URL, cache status,
etc.

In the new engine we were ignoring these events and using the
V8.CompileScript event, but this has no metadata attached to it so is
not useful to us.

This CL updates the new engine to use v8.compile, and updates the UI
code to use it to generate the details. The TimelineUIUtils is a bit
messy and needs breaking down, which we will do as we migrate more of
the code onto exclusively the new engine now that all the tracks are
powered by the new engine.

Bug: 1508029
Change-Id: I259d66f63686f3866e743a08e7b4e32dc6c59746