[Tests] Move animations panel related web tests to E2E tests
These tests use `waitForAnimationAdded` which is changed a bit and instead of changing that I have decided to move the tests to devtools-frontend repo.

There is another test[1] that uses this but it is a not very useful test for us now; so I'll just remove it.

Here are the moved tests:
* http/tests/devtools/animation/animation-timeline.js
* http/tests/devtools/animation/animation-web-anim-negative-start-time.js
* http/tests/devtools/animation/animation-KeyframeEffect-crash.js

[1]: third_party/blink/web_tests/http/tests/devtools/animation/animation-transition-setTiming-crash.js

Bug: 1506561
Change-Id: I2c8eff0db9fd8f36d7ea62ea8f916694b0d2a1ae