Improve device posture support by showing a drop down menu to choose the posture.
On foldable devices it's easier to show the device posture being emulated rather
than a single toggle button.

This patch adds a new emulated device property to distinguish between dual
screens and foldable devices. When a foldable device is selected a drow down
menu will be shown to select the current posture. Selecting a given posture
will then emulate it and set the viewport segments accordingly.

Bug: chromium:1066842
Change-Id: Ia2f4e1c2b2206f37d6ecb6af26f57448af52cc99