[cleanup] Reduce number of View (widget) singletons.
This CL updates `PreRegisteredView` to keep a single widget instance
after loading (until disposing), instead of relying on the `loadView()`
methods to do the deduplication.

This way we can (a) guarantee that we release the resources for view
widget when they are no longer needed and (b) remove a whole lot of
unnecessary singletons for views.

In this CL we remove the now trivially redundant singletons only.
There's much more that is redundant now, but where we need a bit more
work to make that actually happen.

Bug: chromium:1379221
Change-Id: I893a3aaa0883fe65f3f1ad823575dc7671d34f07