Implement and migrate to TraceBounds
Firstly, sorry this is such a huge CL - but it wasn't straightforward to
try to break this down into small pieces.

This CL introduces the use of the TraceBounds service, which stores the
3 relevant bounds we have and allows various components to access them.

As part of this CL I have also included a large README.md with some of
the architectural notes in hof how the timings work, hopefully to help
future developers make sense of it.

This CL is also able to remove a lot of the PerformanceModel code
relating to managing the window, and to slim down the amount of events
we fire and listen to.

Bug: 1491689
Change-Id: I540ca4e1d5f44d897543928a96aa6da929efa642