[ui] Keep the single view instance around forever once created.
This addresses a memory leak introduced by https://crrev.com/c/5104712
and follow up CLs, where we properly create new instances of pre-
registered view widgets when needed, but leak most prior instances due
to event listeners installed on some other global state which are
never removed.

For now, we duck tape the situation by holding on to the single view
instance for the lifetime of the `PreRegisteredView` (so basically
forever).

Bug: chromium:1379221
Change-Id: I22e1eb4b5ff6cb3d552c7f837b0694480333e898