[elements] Improve authoring hint for flex/grid container properties.
For flex/grid container properties like align-content or
justify-content, the current CSS authoring hint is technically
correct, but not very helpful. Instead try to give the developer
a better hint, in particular when it is detected that the parent
is a container and the developer probably should have either put
the declaration onto the parent or used align-self or justify-self
on the child instead.

Screenshot: https://imgur.com/ywW0TZO.png
Fixed: chromium:1394376
Change-Id: I73173ed31f49c4632807beccda2b28e4ff51d3fd