RPP: split out LayoutInvalidation and ScheduleStyle events
For a reason none of us are too sure about, when these were initially
implemented we tied ScheduleStyle events and LayoutInvalidation events
together.

That's not something we want to do generally in the trace engine - so
this CL splits them apart. This is also required to implement an
InvalidationsHandler to migrate the Invalidations tracking code from the
old engine to the new.

Bug: none
Change-Id: I96e3f2c38ccb009bfc7e38062d346a5b93a76243