RPP: track recording time in TimelineController
This is part of the work to get rid of PerformanceModel - next up is the
recordStartTime() method. Unfortunately I cannot remove it fully as the
TimelineHistoryManager relies on it. That is what I am working on next -
but this CL still progresses us by storing the recording time in the
controller and then passing it through, thus ensuring that the
TimelinePanel does not depend on the PerformanceModel to know the time
that the recording started.

Bug: 1499182
Change-Id: I863ec95f8b1e9142e468517cd4bf3c8e83a4110e