RPP: HistoryManager relies on new engine
This CL moves the TimelineHistoryManager to key data based off the index
of that data in the trace engine, rather than using the Performance
Model as the key.

The TimelineHistoryManager still needs to store the perf model, but only
because the TimelinePanel expects it.

Now because the history manager stores items based on the index, I
needed to update some code because code that looked like `if(index)`
would fail when index === 0. I've made these checks stricter to check
for `null` explicitly.

Bug: 1499182
Change-Id: I446aa7ea4a81d8f27e6dd64303191223ca9e35e8