[e2e] Disable two more flaky overrides tests.
In the context of https://crrev.com/c/5116659, these tests became even
more flaky, but primarily when run as part of a test suite run. They
don't tend to be as flaky when run alone (in fact they mostly passed).

I'll disable these for now and work with wolfi@ to improve these.

Bug: chromium:1502463
Change-Id: I6df8c0fdd2451f210cabd65c17adfafea0188803