[application] Re-arrange IndexedDB toolbar buttons.
Move the "Clear all" and "Delete selected" buttons right after the
"Refresh" button, similar to how it's done for the timeline panel.
Also use the `bin` instead of the `close` icon for the "Delete selected"
action. This makes it less likely that one confuses these actions with
the clearing of the input field.

Fixed: chromium:1393800
Screenshot: https://imgur.com/sfHamZe.png
Change-Id: Ie9cf63a8dbd1b7a179425c25a12d370aae078cf2