[RPP] Actually do not redraw minimap when not needed
I noted the check to determine if the minimap needs to be redrawn was
not working properly as I could see the activity happening twice in some
traces. Then I noted that the check included conditions that were
irrelevant to the check because it compared values of variables that
were never set. So I removed these variables completely and checked we
only draw the minimap once per trace.

Bug: 1472155
Change-Id: I3f9a886abd6c4eb1ad86e54bef0217159c567e28