[network] Port Content-Type parsing from blink
This CL ports net::HttpUtil::ParseContentType to TypeScript. We'll use
the frontend parser for:

  1) Importing HAR files. They contain the raw 'Content-Type' header
  2) Parse 'Content-Type' for `Fetch.requestPaused` events as the
     backend does not provide the parsed mime type + charset
  3) Creating requests in the service worker cache view.

These are cases where we don't get the ready mime type/charset from
the backend: Either the backend is not involved (e.g HAR import) or
it does not (yet?) send the information (Fetch.requestPaused).

The accompanying unit tests are also ported over from blink.

Drive-by: Move MimeType enum into separate file together with the
parsing.

R=dsv@chromium.org, wolfi@chromium.org

Bug: 1509336
Change-Id: I6b1435c17d4c1a82fbe158f350d42c07a748b678