RPP: use invalidations handler
This CL uses the new InvalidationsHandler to populate the Invalidations
section of the DetailsView.

This Cl does change how the data is represented - the old view used a
tree view. However, I took the decision to drastically simplify this
because:

1. It is behind an experimental flag, so we can change it more easily.
2. The code that powered the old engine was very hard to follow.
3. I expect this UI to be reworked and changed drastically in the
   future.

Here is a screenshot of what it looks like now: https://imgur.com/a/dsBy2kA

Once this CL lands I can remove the old implementation from the old
engine.

As a drive-by this CL also includes some general tidying up of
TimelineUIUtils.

Bug: 1478975
Change-Id: I127bacf335e0963bcd400f4a7678fa0442291125