[application] Remove confusing "Refresh" button for Origin Trials.
The "Refresh" button was added unnecessarily to begin with, since the
assumption was that the developer would have to reload DevTools itself
in order for new tokens to show up. But that's wrong, since the new
list is fetched every time the component is rendered, so at the very
least each time the developer switches between different panels inside
of the Application panel.

This CL also introduces a proper empty state to indicate that there are
no origin trial tokens, and puts an explanation with a "Learn more" link
into the Origin Trials section.

Screenshot: https://imgur.com/yGwGvyR.png
Fixed: chromium:1315133
Bug: chromium:607555
Change-Id: Ic0c08c4e6c5f951b39ce0f5b2ec6d49d74b1e76d