comfirmDialog box and text popping out of screen while deleting indexedDB fixed
When we try to delete a DB in indexedDB inside storage section in application in inspect element,
the confirm dialog box and its text was popping out of screen when we drag the inspect element to
right increasing the page view.


Updated some css in confirmDialog.css file to fix the issue.

Change-Id: Icbb35980e613a382ef6ae6cb47e25cd9bf15bb18