[DevToolsIcons] Merge `NewIcon` into `Icon` component.
Change the `Icon` component to use CSS for dimensions and color (based
on what we did for `NewIcon`), and keep the `data` property as a legacy
way to set all at once.

The `Icon` now has an observed "name" attribute (reflected as `name`
property), which accepts both name and path.

This also adds a bunch of tests for the `Icon` class (and improves the
tests for other classes using it).

Drive-by-fix: Simplify the `RequestLinkIcon` and `IssueLinkIcon` classes
and their tests by using CSS to style the icon and not relying on the
`RenderCoordinator` being triggered by the `Icon` component.

Bug: chromium:1421730, chromium:1514825
Change-Id: I067faf9fcef1f02092b369ee0c59776e4674ee6e
Fixed: chromium:1405244