Fixed to display in devtools when HTTP Status Code is 0
There is a bug where HTTP Status Code is not displayed in Network Tab of
devtools when HTTP Status Code is 0.

Therefore, I changed the initial value of HTTP Status Code from 0 to
null and fixed the conditional branching of the display. This has been
fixed so that the HTTP Status Code is displayed correctly.

This issue also mentions negative HTTP Status Codes, but this patch does
not fix it.

Bug: 1351769
Change-Id: I997314e085a5b195eb6439003ef9df88a4e79e2b