[ssci] Update Lighthouse metadata
This CL updates the URLs for Lighthouse and Lighthouse i18n.js, so that
it is an absolute URL and successfully links in chrome://credits
instead of being a relative link.

Also specified the Version for Lighthouse, as at least one of Version,
Revision or Date should be specified for third party metadata.

Bug: b:290296624
Change-Id: Ib31c5e9ad5f555a0100ebf3591b1c094aaf80294