[DevToolsIcons] Migrate `GlassPane` off of legacy `Icon`.
This removes the use of the legacy `Icon` component from the `GlassPane`
and instead goes for a pure CSS solution, since this is not a suitable
use case for the new `Icon` component (it's not even actually an icon,
but the arrow of the glass pane). We'll need to decide later if this is
the desired long-term solution, or if we want to go with a different
approach instead.

Bug: chromium:1421730
Change-Id: I515130218591e8e8e3bfcdb8070dd53e2d545822