[sources] Tweak the Watch expressions view a bit.
Migrate off of the legacy `Icon` component and use a proper `Button`
instead with the trash icon for the delete action. Also get rid of the
resizer left/right align logic, since we now have a minimum width
set for the Sources panel sidebar anyways.

Also fix a typo in the CSS where it set the `min-height` instead of
the `min-width` for the name.

Bug: chromium:1514825, chromium:1421730, chromium:1292637
Change-Id: I670d3d3c97a4b9e0f6cbb2b17d10d254c5a0d6e0